backend: prices can be changed

+++ backend/src/main/java/de/fasterfood/fasterfood/ingredient/Ingredient.java
@@ -63,7 +63,7 @@ public class Ingredient {
         return purchasePrice;
     }
 
-    public void setPurchasePrice(int purchasePrice) {
+    public void setPurchasePrice(double purchasePrice) {
         this.purchasePrice = purchasePrice;
     }
